The app's ui now has its own activity.
